Add files via upload
This is a basic data entry built in Microsoft Excel using VBA Programming. The user will be ask to input the personal information that will be automate to put in the ID (Identification). It has a upload feature where the user can upload its 1x1 picture from the files directly. It has also a print button where the finished ID will automatically be download into PDF